requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.3
attrs==20.3.0
Click==7.0
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.1
django-annoying==0.10.6
django-etc==1.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-jsonstore==0.4.1
django-material==1.7.1
django-viewflow==1.6.1
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.0
pefile==2019.4.18
Pillow==6.2.1
psycopg2==2.8.6
psycopg2-binary==2.8.4
PyInstaller==3.6
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
six==1.13.0
sqlparse==0.3.0
virtualenv==20.1.0
Werkzeug==0.16.0
whitenoise==5.2.0
